[PATCH] remove unnecessary PAE pgd set
From: Dave Hansen <haveblue@us.ibm.com>, Bill Irwin

With PAE on, there are only 4 PGD entries.  The kernel ones never change,
so there is no need to copy them when a vmalloc fault occurs.  This was
this was causing problems with the split pmd patches, but it is still
correct for mainline.

Tested with and without PAE.  I ran it in a loop turning on and off 10 swap
partitions, which is what excited the original bug.
http://bugme.osdl.org/show_bug.cgi?id=640