[PATCH] scsi: allow devices to restrict start on add
When a SCSI disk is added and it returns a NOT READY the SD driver is 
automatically sending a START_UNIT command to spin the device up.  While this 
may be the desired behavior for many if not most devices not all devices 
either want or need this.  The attached patch provides a mechanism via the 
device_list that allows a device to be defined to disable the automatic start 
being issued on an add.

The patch also modifies the device_list for several devices that would prefer 
to not have the start command issued.