[PATCH] SCSI: Make sysfs attributes mutable
The attached patch makes the sysfs attributes a property of the host
template (I can't really see a reason why the actual attributes need to
change per host, so I think the template is the better place for them).

I've also provided helper functions to modify the attributes in the LLD
init routines.