[PATCH] make common m68knommu/68328 specific ints.c
Modify the m68knommu/68328 specific ints.c to be the general ints
handler when building for this sub-architecture. It is just simpler to
have one for each sub-architecture (which means we currently need 3
for the 3 prominant m68knommu families). Each can handle the hardware
setup differences, and there is a few at this level. This doesn't
really add much code overall, since 2 of the 3 m68knommu architectures
already had significant specific int handling code.