[PATCH] i2o memleak comment
From: Andy Whitcroft <apw@shadowen.org>

There's a spot in i2o where we deliberately leak some memory when the
hardware plays up.  The alternative is to let the hardware scribble on it at
some unknown time in the future.

Things like the Stanford checker keep alleging that this is a bug.  So shut
them up with a comment