[PATCH] get rid of CONFIG_ALL_PPC
This patch gets rid of CONFIG_ALL_PPC, which was a very confusing
option, since it didn't actually mean "ALL" at all, it was more a
"common set" thing.

The primary replacement for CONFIG_ALL_PPC is CONFIG_PPC_MULTIPLATFORM.
I have also defined CONFIG_PPC_PMAC, CONFIG_PPC_PREP and CONFIG_PPC_CHRP
for selecting code which is only needed for one of the three platforms
that CONFIG_ALL_PPC represented.  This is something that we (the PPC
community) have been talking about doing for some time.  There is also a
CONFIG_PPC_OF which is for PPC machines with Open Firmware, which is
currently powermacs and CHRP machines.

At the moment, CONFIG_PPC_{PMAC,PREP,CHRP,OF} get unconditionally
defined if CONFIG_PPC_MULTIPLATFORM is selected, but in future this
split will let us have more control over what gets included, so that
for example we don't necessarily have to include powermac bits in a
kernel for a PReP machine.

I have gone through the uses of CONFIG_ALL_PPC one by one and decided
which of the new symbols best represents the set of machines that need
the code in question.  In fact most of the uses of CONFIG_ALL_PPC in
the drivers have been replaced by CONFIG_PPC_PMAC.  The other symbols
are mostly confined to the PPC architecture code.