[fs] Remove kobject support for filesystems
It was initially added for the immediate gain of being able to see what 
filesystems were registered in the system. However, it was done without
proper analysis of the lifetime rules associated with them. 

Someday, we should convert struct filesystem_type and struct super_block to
use kobjects, though it will take sufficent time and effort to get it right.
Until then, we go without..