[PATCH] typo in new class_device_release
There is a typo in the current 2.5.70 bk version of class_device_release that
was not there in my original patch.  By confusing the class and the class_device,
the release function oops.  cd->release is always the function itself (class_device_release),
cls->release is the one setup for the class (net_class in my case).